refactor: Modal 에 사용되는 constant 파일 위치 변경
