+++ components/overlay/modal/LimitExceededErrorModal.tsx
@@ -1,5 +1,5 @@
 import AlertModal from "@/components/overlay/modal/AlertModal";
-import { LOGIN_ERROR_MESSAGE } from "@/constants/ModalMessages";
+import { LOGIN_ERROR_MESSAGE } from "@/constants/modal/ModalMessages";
 
 const LimitExceededErrorModal = () => {
   const { title, content } = LOGIN_ERROR_MESSAGE.LIMIT_EXCEEDED;
